+++ sources/ft_game/ft_wall_texture.c
@@ -6,7 +6,7 @@
 /*   By: tlassere <tlassere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/19 22:08:13 by tlassere          #+#    #+#             */
-/*   Updated: 2024/05/01 17:24:59 by tlassere         ###   ########.fr       */
+/*   Updated: 2024/05/02 15:16:08 by tlassere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ size_t	ft_get_y_pos(long double ratio, long double pos_y,
  * @param pos_x value 0 >= && .9999... <= value; to get X of image
  * @return ptr of pixel in X
 */
-static uint8_t	*ft_get_x_pos(mlx_texture_t *texture, float pos_x)
+static uint8_t	*ft_get_x_pos(mlx_texture_t *texture, double pos_x)
 {
 	return (texture->pixels + (size_t)(texture->width * pos_x)
 		* sizeof(int));
@@ -57,7 +57,7 @@ void	ft_init_texture(t_texture_pos *texture_pos,
 }
 
 void	ft_print_line_texture(t_data *data, mlx_texture_t *texture,
-	t_data_wall wall, float pos_x)
+	t_data_wall wall, double pos_x)
 {
 	t_texture_pos	texture_pos;
 	size_t			texture_width;
